Rename clsuterformation -> testclusters
