--- buildSrc/src/main/resources/META-INF/gradle-plugins/elasticsearch.clusterformation.properties
@@ -1,1 +1,0 @@
-implementation-class=org.elasticsearch.gradle.clusterformation.ClusterformationPlugin
