+++ buildSrc/src/main/resources/META-INF/gradle-plugins/elasticsearch.testclusters.properties
@@ -1,0 +1,1 @@
+implementation-class=org.elasticsearch.gradle.testclusters.TestClustersPlugin
